moves some stuff around
