+++ tests/test_RetrySender.js
@@ -1,9 +1,9 @@
 const chai = require("chai");
 const expect = chai.expect;
 const RetrySender = require("../src/RetrySender");
-const FailingSender = require("./fixtures/FailingSender");
+const {MockSenderWithStatusCodesAndHeaders} = require("./fixtures/mock_senders");
 const Request = require("../src/Request.js");
-const FakeSleeper = require("../src/util/FakeSleeper");
+const FakeSleeper = require("./fixtures/FakeSleeper");
 
 async function sendWithRetry(retries, inner, sleeper) {
 	const request = new Request();
@@ -13,28 +13,28 @@ async function sendWithRetry(retries, inner, sleeper) {
 
 describe ("Retry Sender tests", function () {
 	it("test success does not retry", async function () {
-		let inner = new FailingSender(["200"]);
+		let inner = new MockSenderWithStatusCodesAndHeaders(["200"]);
 		await sendWithRetry(5, inner, new FakeSleeper());
 
 		expect(inner.currentStatusCodeIndex).to.equal(1);
 	});
 
 	it("test client error does not retry", async function () {
-		let inner = new FailingSender(["422"]);
+		let inner = new MockSenderWithStatusCodesAndHeaders(["422"]);
 		await sendWithRetry(5, inner, new FakeSleeper());
 
 		expect(inner.currentStatusCodeIndex).to.equal(1);
 	});
 
 	it("test will retry until success", async function () {
-		let inner =  new FailingSender(["500", "500", "500", "200", "500"]);
+		let inner =  new MockSenderWithStatusCodesAndHeaders(["500", "500", "500", "200", "500"]);
 		await sendWithRetry(10, inner, new FakeSleeper());
 
 		expect(inner.currentStatusCodeIndex).to.equal(4);
 	});
 
 	it("test return response if retry limit exceeded", async function () {
-		let inner = new FailingSender(["500", "500", "500", "500", "500"]);
+		let inner = new MockSenderWithStatusCodesAndHeaders(["500", "500", "500", "500", "500"]);
 		const sleeper = new FakeSleeper();
 		const response = await sendWithRetry(4, inner, sleeper);
 
@@ -45,7 +45,7 @@ describe ("Retry Sender tests", function () {
 	});
 
 	it("test backoff does not exceed max", async function () {
-		let inner = new FailingSender(["500", "500", "500", "500", "500", "500", "500", "500", "500", "500", "500", "500", "500", "200"]);
+		let inner = new MockSenderWithStatusCodesAndHeaders(["500", "500", "500", "500", "500", "500", "500", "500", "500", "500", "500", "500", "500", "200"]);
 		const sleeper = new FakeSleeper();
 
 		await sendWithRetry(20, inner, sleeper);
@@ -54,14 +54,14 @@ describe ("Retry Sender tests", function () {
 	});
 
 	it("test empty status does not retry", async function () {
-		let inner = new FailingSender([]);
+		let inner = new MockSenderWithStatusCodesAndHeaders([]);
 		await sendWithRetry(5, inner, new FakeSleeper());
 
 		expect(inner.currentStatusCodeIndex).to.equal(1);
 	});
 
 	it("test sleep on rate limit", async function () {
-		let inner = new FailingSender(["429", "200"]);
+		let inner = new MockSenderWithStatusCodesAndHeaders(["429", "200"]);
 		const sleeper = new FakeSleeper();
 
 		await sendWithRetry(5, inner, sleeper);
@@ -70,7 +70,7 @@ describe ("Retry Sender tests", function () {
 	});
 
 	it("test rate limit error return", async function () {
-		let inner = new FailingSender(["429"], {"Retry-After": 7});
+		let inner = new MockSenderWithStatusCodesAndHeaders(["429"], {"Retry-After": 7});
 		const sleeper = new FakeSleeper();
 
 		await sendWithRetry(10, inner, sleeper);
@@ -79,7 +79,7 @@ describe ("Retry Sender tests", function () {
 	});
 
 	it("test retry after invalid value", async function () {
-		let inner = new FailingSender(["429"], {"Retry-After": "a"});
+		let inner = new MockSenderWithStatusCodesAndHeaders(["429"], {"Retry-After": "a"});
 		const sleeper = new FakeSleeper();
 
 		await sendWithRetry(10, inner, sleeper);
@@ -88,7 +88,7 @@ describe ("Retry Sender tests", function () {
 	});
 
 	it("test retry error", async function () {
-		let inner = new FailingSender(["429"], undefined, "Big Bad");
+		let inner = new MockSenderWithStatusCodesAndHeaders(["429"], undefined, "Big Bad");
 		const sleeper = new FakeSleeper();
 
 		const response = await sendWithRetry(10, inner, sleeper);
